refactored file structure based on tony's request, modified the card components to be bigger based on our agreement
